Moved composite pk None lookup test into composite_pk app and asserted queryset results.
